# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/health": {
      "get": {
        "operationId": "getHealth",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponseDto"
                }
              }
            }
          }
        },
        "summary": "Liveness probe",
        "tags": [
          "health"
        ]
      }
    },
    "/api/announcements": {
      "get": {
        "description": "Supports full-text search over title and body, ANY-match category filtering and pagination. Always ordered by lastUpdate descending.",
        "operationId": "listAnnouncements",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Case-insensitive match against title or body.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryIds",
            "required": false,
            "in": "query",
            "description": "ANY-match: announcements having at least one of these categories.",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "format": "uuid"
              }
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnouncementListResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "List announcements",
        "tags": [
          "announcements"
        ]
      },
      "post": {
        "description": "Requires at least one existing category. lastUpdate is server-managed and cannot be supplied.",
        "operationId": "createAnnouncement",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAnnouncementDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnouncementResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or a referenced category does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create an announcement",
        "tags": [
          "announcements"
        ]
      }
    },
    "/api/announcements/{id}": {
      "get": {
        "operationId": "getAnnouncement",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnouncementResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Malformed id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Announcement does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get a single announcement",
        "tags": [
          "announcements"
        ]
      },
      "patch": {
        "description": "Partial update. Providing categoryIds replaces the whole category set; every update refreshes lastUpdate.",
        "operationId": "updateAnnouncement",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAnnouncementDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnouncementResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or a referenced category does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Announcement does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Update an announcement",
        "tags": [
          "announcements"
        ]
      },
      "delete": {
        "operationId": "deleteAnnouncement",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Announcement deleted"
          },
          "404": {
            "description": "Announcement does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Delete an announcement",
        "tags": [
          "announcements"
        ]
      }
    },
    "/api/categories": {
      "get": {
        "operationId": "listCategories",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "List all categories alphabetically",
        "tags": [
          "categories"
        ]
      },
      "post": {
        "description": "Category names are unique case-insensitively: \"Health\", \"health\" and \"  HEALTH  \" are considered the same category.",
        "operationId": "createCategory",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "A category with the same normalized name already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create a category",
        "tags": [
          "categories"
        ]
      }
    }
  },
  "info": {
    "title": "Announcement Manager API",
    "description": "REST API for managing announcements and their categories.",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "health",
      "description": "Liveness"
    },
    {
      "name": "announcements",
      "description": "Announcement management"
    },
    {
      "name": "categories",
      "description": "Category management"
    }
  ],
  "servers": [],
  "components": {
    "schemas": {
      "HealthResponseDto": {
        "type": "object",
        "properties": {}
      },
      "CategoryResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "createdAt",
          "updatedAt"
        ]
      },
      "AnnouncementResponseDto": {
        "type": "object",
        "properties": {
          "publicationDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastUpdate": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryResponseDto"
            }
          }
        },
        "required": [
          "publicationDate",
          "lastUpdate",
          "createdAt",
          "id",
          "title",
          "body",
          "categories"
        ]
      },
      "AnnouncementListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AnnouncementResponseDto"
            }
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "total": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "page",
          "limit",
          "total"
        ]
      },
      "ApiErrorResponseDto": {
        "type": "object",
        "properties": {
          "code": {
            "description": "Stable machine-readable error code; frontend logic branches on this.",
            "enum": [
              "VALIDATION_ERROR",
              "ANNOUNCEMENT_NOT_FOUND",
              "CATEGORY_NOT_FOUND",
              "CATEGORY_ALREADY_EXISTS",
              "INVALID_CATEGORY_SELECTION",
              "SEARCH_UNAVAILABLE",
              "NOT_FOUND",
              "INTERNAL_ERROR"
            ],
            "type": "string"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "description": "Present on validation failures: field name -> list of problems.",
            "example": {
              "title": [
                "Title is required"
              ]
            }
          },
          "statusCode": {
            "type": "number",
            "description": "HTTP status code, duplicated in the body for logging convenience."
          },
          "message": {
            "type": "string",
            "description": "Human-readable summary of the problem."
          }
        },
        "required": [
          "code",
          "statusCode",
          "message"
        ]
      },
      "CreateAnnouncementDto": {
        "type": "object",
        "properties": {
          "publicationDate": {
            "type": "string",
            "description": "ISO 8601 datetime; the client converts from its local input format.",
            "format": "date-time"
          },
          "title": {
            "type": "string",
            "maxLength": 200
          },
          "body": {
            "type": "string"
          },
          "categoryIds": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "publicationDate",
          "title",
          "body",
          "categoryIds"
        ]
      },
      "UpdateAnnouncementDto": {
        "type": "object",
        "properties": {
          "publicationDate": {
            "type": "string",
            "format": "date-time"
          },
          "title": {
            "type": "string",
            "maxLength": 200
          },
          "body": {
            "type": "string"
          },
          "categoryIds": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100
          }
        },
        "required": [
          "name"
        ]
      }
    }
  }
}